refactor(steam-id-form): move hints and validation to external files
